Comment cleaning and bug fixing

+++ app.py
@@ -1,13 +1,13 @@
-from flask import Flask # Para usar flask
-from flask import render_template # Para redireccionar una u otra URL
-from flask import request # Para tomar datos de formularios etc.
-from flask import redirect # Para redirigir a la misma pagina (o a otra? no se)
+from flask import Flask
+from flask import render_template
+from flask import request
+from flask import redirect
 from flask import send_from_directory
-from flaskext.mysql import MySQL # Para que Flask pueda conectarse con sql
-from datetime import datetime # Para tomar la fecha y hora de la computadora, la usamos para cambiar el nombre de archivo de la imagen
+from flaskext.mysql import MySQL
+from datetime import datetime
 import os
 
-app = Flask(__name__) # Siempre se arranca asi
+app = Flask(__name__)
 
 mysql = MySQL()
 app.config['MYSQL_DATABASE_HOST'] = 'localhost'
@@ -16,32 +16,30 @@ app.config['MYSQL_DATABASE_PASSWORD'] = ''
 app.config['MYSQL_DATABASE_DB'] = 'sistema22072'
 mysql.init_app(app)
 
-CARPETA = os.path.join('uploads') # Se ingresa la ruta del directorio que quiero referenciar
-app.config['CARPETA'] = CARPETA # Cuando en el sistema hago referencia a la palabra CARPETA, hago referencia al directorio uploads
+CARPETA = os.path.join('uploads')
+app.config['CARPETA'] = CARPETA
 
-@app.route('/uploads/<nombreFoto>') # Cuando alguien quiere levantar una foto, el codigo lo va a reconocer aca y va a ejecutar el codigo
+@app.route('/uploads/<nombreFoto>')
 def uploads(nombreFoto):
     return send_from_directory(app.config['CARPETA'], nombreFoto)
 
-@app.route('/') # Lo que sucede al entrar al index
-def index():  # ese 'index' puede tener otro nombre que yo quiera.
-    sql =  "SELECT * FROM `empleados`;" # Quiero que el index me muestre todo lo que esta cargado.
-    conn = mysql.connect() # Se abre la coneccion
-    cursor = conn.cursor() # Se crea un cursor
-    cursor.execute(sql) # Hago que el cursor ejecute la consulta
-    empleados = cursor.fetchall() # Trae todos los datos y los mete en una tupla
-    conn.commit() # Se envian los cambios a la base de datos.
-    return render_template('empleados/index.html', empleados=empleados) # En el segundo parametro le pongo el nombre de la variable que quiero
+@app.route('/')
+def index():
+    sql =  "SELECT * FROM `empleados`;"
+    conn = mysql.connect()
+    cursor = conn.cursor()
+    cursor.execute(sql)
+    empleados = cursor.fetchall()
+    conn.commit()
+    return render_template('empleados/index.html', empleados=empleados)
 
 @app.route("/destroy/<int:id>")
 def destroy(id):
     conn = mysql.connect()
     cursor = conn.cursor()
 
-    cursor.execute("SELECT foto FROM empleados WHERE id=%s", id) # Se selecciona el empleado
-    #fila = cursor.fetchall() # Se obtienen los datos
-    #os.remove(os.path.join(app.config['CARPETA'], fila[0][0])) # Se elimina la imagen guardada en la carpeta uploads
-    fila = cursor.fetchone() # Se obtienen los datos. fetchone obtiene un solo dato, por lo que es bueno usarlo cuando el id es unico. Ver que en otra uso fetchall, eso trae mas datos y hay que poner fila[0][0]
+    cursor.execute("SELECT foto FROM empleados WHERE id=%s", id)
+    fila = cursor.fetchone()
 
     if fila[0] != "":
         os.remove(os.path.join(app.config['CARPETA'], fila[0]))
@@ -54,14 +52,14 @@ def destroy(id):
 def edit(id):
     conn = mysql.connect()
     cursor = conn.cursor()
-    cursor.execute("SELECT * FROM empleados WHERE id=%s", id) # Solo se busca el empleado con el id que se ingreso
+    cursor.execute("SELECT * FROM empleados WHERE id=%s", id)
     empleados = cursor.fetchall()
     conn.commit()
-    return render_template('empleados/edit.html', empleados=empleados) # Envia al usuario a /edit
+    return render_template('empleados/edit.html', empleados=empleados)
 
 @app.route("/update", methods=['POST'])
 def update():
-    id = request.form['txtId'] # Para poder usar el id aca, HAY que ponerlo en el html. Despues se puede ocultar.
+    id = request.form['txtId']
     _nombre = request.form['txtNombre']
     _correo = request.form['txtCorreo']
     _foto = request.files['txtFoto']
@@ -77,22 +75,24 @@ def update():
     if _foto.filename != '':
         nuevoNombreFoto = tiempo + _foto.filename
         _foto.save("uploads/" + nuevoNombreFoto)
-
+        
         cursor.execute("SELECT foto FROM empleados WHERE id=%s", id)
-        fila = cursor.fetchall() # Todo esto es para conseguir el nombre de la foto anterior (que se quiere eliminar ya que se reemplaza con otra)
-        os.remove(os.path.join(app.config['CARPETA'], fila[0][0]))
-        # remove agarra dos parametros. Primero a donde se quiere entrar, y despues que se quiere borrar
+        fila = cursor.fetchall()
+
+        if fila[0][0] != "":
+            os.remove(os.path.join(app.config['CARPETA'], fila[0][0]))
+
         cursor.execute("UPDATE empleados SET foto=%s WHERE id=%s", (nuevoNombreFoto, id))
 
     cursor.execute(sql, datos)
     conn.commit()
     return redirect('/')
 
-@app.route("/create") # Al entrar a /create, se muestra el contenido de create.html
+@app.route("/create")
 def create():
     return render_template('empleados/create.html')
 
-@app.route("/store", methods=['POST']) # Al entrar a /store, se toman los datos del formulario
+@app.route("/store", methods=['POST'])
 def storage():
     _nombre = request.form['txtNombre']
     _correo = request.form['txtCorreo']
@@ -102,22 +102,19 @@ def storage():
     tiempo = now.strftime("%Y%H%M%S")
 
     nuevoNombreFoto = ''
-    if _foto.filename != '': # Se verifica que la foto se haya cargado. Si no, nombre es vacio.
+    if _foto.filename != '':
         nuevoNombreFoto = tiempo + _foto.filename
         _foto.save("uploads/" + nuevoNombreFoto)
-    else:
-        nuevoNombreFoto = ""
 
     sql =  "INSERT INTO `empleados` (`id`, `nombre`, `correo`, `foto`) VALUES (NULL, %s, %s, %s);"
     
     datos = (_nombre, _correo, nuevoNombreFoto)
     
-    # Carga en la BD (Base de Datos) VVVV
     conn = mysql.connect()
     cursor = conn.cursor()
-    cursor.execute(sql, datos) # Se ejecuta la consulta en mysql, y el segundo parametro es la lista de datos con los que se reemplazan %s del string
+    cursor.execute(sql, datos)
     conn.commit()
-    return redirect('/') # Redirige a la app a @app.route('/') y hace todo lo que está ahi (busca en la BD y carga eso en el index)
+    return redirect('/')
 
 if __name__ == '__main__':
-    app.run(debug=True) # Ejecuta la aplicacion y comienza a correr el controlador
+    app.run(debug=True)
